LayoutTests/http/tests/w3c/dom/nodes/Document-characterSet-normalization-expected.txt


FAIL Name "utf-8" has label "unicode-1-1-utf-8" (characterSet) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "unicode-1-1-utf-8" (inputEncoding) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf-8" (characterSet) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf-8" (inputEncoding) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf8" (characterSet) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf8" (inputEncoding) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf-16" (characterSet) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf-16" (inputEncoding) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf-16le" (characterSet) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf-16le" (inputEncoding) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf-16be" (characterSet) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "utf-8" has label "utf-16be" (inputEncoding) assert_equals: expected "UTF-8" but got "ISO-8859-1"
FAIL Name "ibm866" has label "866" (characterSet) assert_equals: expected "IBM866" but got "ISO-8859-1"
FAIL Name "ibm866" has label "866" (inputEncoding) assert_equals: expected "IBM866" but got "ISO-8859-1"
FAIL Name "ibm866" has label "cp866" (characterSet) assert_equals: expected "IBM866" but got "ISO-8859-1"
FAIL Name "ibm866" has label "cp866" (inputEncoding) assert_equals: expected "IBM866" but got "ISO-8859-1"
FAIL Name "ibm866" has label "csibm866" (characterSet) assert_equals: expected "IBM866" but got "ISO-8859-1"
FAIL Name "ibm866" has label "csibm866" (inputEncoding) assert_equals: expected "IBM866" but got "ISO-8859-1"
FAIL Name "ibm866" has label "ibm866" (characterSet) assert_equals: expected "IBM866" but got "ISO-8859-1"
FAIL Name "ibm866" has label "ibm866" (inputEncoding) assert_equals: expected "IBM866" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "csisolatin2" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "csisolatin2" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso-8859-2" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso-8859-2" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso-ir-101" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso-ir-101" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso8859-2" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso8859-2" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso88592" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso88592" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso_8859-2" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso_8859-2" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso_8859-2:1987" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "iso_8859-2:1987" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "l2" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "l2" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "latin2" (characterSet) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-2" has label "latin2" (inputEncoding) assert_equals: expected "ISO-8859-2" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "csisolatin3" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "csisolatin3" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso-8859-3" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso-8859-3" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso-ir-109" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso-ir-109" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso8859-3" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso8859-3" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso88593" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso88593" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso_8859-3" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso_8859-3" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso_8859-3:1988" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "iso_8859-3:1988" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "l3" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "l3" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "latin3" (characterSet) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-3" has label "latin3" (inputEncoding) assert_equals: expected "ISO-8859-3" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "csisolatin4" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "csisolatin4" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso-8859-4" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso-8859-4" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso-ir-110" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso-ir-110" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso8859-4" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso8859-4" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso88594" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso88594" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso_8859-4" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso_8859-4" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso_8859-4:1988" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "iso_8859-4:1988" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "l4" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "l4" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "latin4" (characterSet) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-4" has label "latin4" (inputEncoding) assert_equals: expected "ISO-8859-4" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "csisolatincyrillic" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "csisolatincyrillic" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "cyrillic" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "cyrillic" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso-8859-5" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso-8859-5" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso-ir-144" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso-ir-144" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso8859-5" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso8859-5" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso88595" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso88595" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso_8859-5" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso_8859-5" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso_8859-5:1988" (characterSet) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-5" has label "iso_8859-5:1988" (inputEncoding) assert_equals: expected "ISO-8859-5" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "arabic" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "arabic" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "asmo-708" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "asmo-708" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596e" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596e" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596i" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csiso88596i" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csisolatinarabic" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "csisolatinarabic" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "ecma-114" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "ecma-114" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso-8859-6" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso-8859-6" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso-8859-6-e" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso-8859-6-e" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso-8859-6-i" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso-8859-6-i" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso-ir-127" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso-ir-127" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso8859-6" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso8859-6" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso88596" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso88596" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso_8859-6" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso_8859-6" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso_8859-6:1987" (characterSet) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-6" has label "iso_8859-6:1987" (inputEncoding) assert_equals: expected "ISO-8859-6" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "csisolatingreek" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "csisolatingreek" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "ecma-118" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "ecma-118" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "elot_928" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "elot_928" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "greek" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "greek" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "greek8" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "greek8" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso-8859-7" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso-8859-7" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso-ir-126" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso-ir-126" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso8859-7" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso8859-7" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso88597" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso88597" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso_8859-7" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso_8859-7" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso_8859-7:1987" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "iso_8859-7:1987" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "sun_eu_greek" (characterSet) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-7" has label "sun_eu_greek" (inputEncoding) assert_equals: expected "ISO-8859-7" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "csiso88598e" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "csiso88598e" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "csisolatinhebrew" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "csisolatinhebrew" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "hebrew" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "hebrew" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso-8859-8" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso-8859-8" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso-8859-8-e" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso-8859-8-e" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso-ir-138" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso-ir-138" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso8859-8" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso8859-8" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso88598" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso88598" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso_8859-8" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso_8859-8" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso_8859-8:1988" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "iso_8859-8:1988" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "visual" (characterSet) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8" has label "visual" (inputEncoding) assert_equals: expected "ISO-8859-8" but got "ISO-8859-1"
FAIL Name "iso-8859-8-i" has label "csiso88598i" (characterSet) assert_equals: expected "ISO-8859-8-I" but got "ISO-8859-1"
FAIL Name "iso-8859-8-i" has label "csiso88598i" (inputEncoding) assert_equals: expected "ISO-8859-8-I" but got "ISO-8859-1"
FAIL Name "iso-8859-8-i" has label "iso-8859-8-i" (characterSet) assert_equals: expected "ISO-8859-8-I" but got "ISO-8859-1"
FAIL Name "iso-8859-8-i" has label "iso-8859-8-i" (inputEncoding) assert_equals: expected "ISO-8859-8-I" but got "ISO-8859-1"
FAIL Name "iso-8859-8-i" has label "logical" (characterSet) assert_equals: expected "ISO-8859-8-I" but got "ISO-8859-1"
FAIL Name "iso-8859-8-i" has label "logical" (inputEncoding) assert_equals: expected "ISO-8859-8-I" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "csisolatin6" (characterSet) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "csisolatin6" (inputEncoding) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso-8859-10" (characterSet) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso-8859-10" (inputEncoding) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso-ir-157" (characterSet) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso-ir-157" (inputEncoding) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso8859-10" (characterSet) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso8859-10" (inputEncoding) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso885910" (characterSet) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "iso885910" (inputEncoding) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "l6" (characterSet) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "l6" (inputEncoding) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "latin6" (characterSet) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-10" has label "latin6" (inputEncoding) assert_equals: expected "ISO-8859-10" but got "ISO-8859-1"
FAIL Name "iso-8859-13" has label "iso-8859-13" (characterSet) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
FAIL Name "iso-8859-13" has label "iso-8859-13" (inputEncoding) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
FAIL Name "iso-8859-13" has label "iso8859-13" (characterSet) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
FAIL Name "iso-8859-13" has label "iso8859-13" (inputEncoding) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
FAIL Name "iso-8859-13" has label "iso885913" (characterSet) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
FAIL Name "iso-8859-13" has label "iso885913" (inputEncoding) assert_equals: expected "ISO-8859-13" but got "ISO-8859-1"
FAIL Name "iso-8859-14" has label "iso-8859-14" (characterSet) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
FAIL Name "iso-8859-14" has label "iso-8859-14" (inputEncoding) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
FAIL Name "iso-8859-14" has label "iso8859-14" (characterSet) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
FAIL Name "iso-8859-14" has label "iso8859-14" (inputEncoding) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
FAIL Name "iso-8859-14" has label "iso885914" (characterSet) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
FAIL Name "iso-8859-14" has label "iso885914" (inputEncoding) assert_equals: expected "ISO-8859-14" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "csisolatin9" (characterSet) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "csisolatin9" (inputEncoding) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso-8859-15" (characterSet) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso-8859-15" (inputEncoding) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso8859-15" (characterSet) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso8859-15" (inputEncoding) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso885915" (characterSet) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso885915" (inputEncoding) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso_8859-15" (characterSet) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "iso_8859-15" (inputEncoding) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "l9" (characterSet) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-15" has label "l9" (inputEncoding) assert_equals: expected "ISO-8859-15" but got "ISO-8859-1"
FAIL Name "iso-8859-16" has label "iso-8859-16" (characterSet) assert_equals: expected "ISO-8859-16" but got "ISO-8859-1"
FAIL Name "iso-8859-16" has label "iso-8859-16" (inputEncoding) assert_equals: expected "ISO-8859-16" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "cskoi8r" (characterSet) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "cskoi8r" (inputEncoding) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi" (characterSet) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi" (inputEncoding) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi8" (characterSet) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi8" (inputEncoding) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi8-r" (characterSet) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi8-r" (inputEncoding) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi8_r" (characterSet) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-r" has label "koi8_r" (inputEncoding) assert_equals: expected "KOI8-R" but got "ISO-8859-1"
FAIL Name "koi8-u" has label "koi8-u" (characterSet) assert_equals: expected "KOI8-U" but got "ISO-8859-1"
FAIL Name "koi8-u" has label "koi8-u" (inputEncoding) assert_equals: expected "KOI8-U" but got "ISO-8859-1"
FAIL Name "macintosh" has label "csmacintosh" (characterSet) assert_equals: expected "MACINTOSH" but got "ISO-8859-1"
FAIL Name "macintosh" has label "csmacintosh" (inputEncoding) assert_equals: expected "MACINTOSH" but got "ISO-8859-1"
FAIL Name "macintosh" has label "mac" (characterSet) assert_equals: expected "MACINTOSH" but got "ISO-8859-1"
FAIL Name "macintosh" has label "mac" (inputEncoding) assert_equals: expected "MACINTOSH" but got "ISO-8859-1"
FAIL Name "macintosh" has label "macintosh" (characterSet) assert_equals: expected "MACINTOSH" but got "ISO-8859-1"
FAIL Name "macintosh" has label "macintosh" (inputEncoding) assert_equals: expected "MACINTOSH" but got "ISO-8859-1"
FAIL Name "macintosh" has label "x-mac-roman" (characterSet) assert_equals: expected "MACINTOSH" but got "ISO-8859-1"
FAIL Name "macintosh" has label "x-mac-roman" (inputEncoding) assert_equals: expected "MACINTOSH" but got "ISO-8859-1"
FAIL Name "windows-874" has label "dos-874" (characterSet) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "dos-874" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "iso-8859-11" (characterSet) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "iso-8859-11" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "iso8859-11" (characterSet) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "iso8859-11" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "iso885911" (characterSet) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "iso885911" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "tis-620" (characterSet) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "tis-620" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "windows-874" (characterSet) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-874" has label "windows-874" (inputEncoding) assert_equals: expected "WINDOWS-874" but got "ISO-8859-1"
FAIL Name "windows-1250" has label "cp1250" (characterSet) assert_equals: expected "WINDOWS-1250" but got "ISO-8859-1"
FAIL Name "windows-1250" has label "cp1250" (inputEncoding) assert_equals: expected "WINDOWS-1250" but got "ISO-8859-1"
FAIL Name "windows-1250" has label "windows-1250" (characterSet) assert_equals: expected "WINDOWS-1250" but got "ISO-8859-1"
FAIL Name "windows-1250" has label "windows-1250" (inputEncoding) assert_equals: expected "WINDOWS-1250" but got "ISO-8859-1"
FAIL Name "windows-1250" has label "x-cp1250" (characterSet) assert_equals: expected "WINDOWS-1250" but got "ISO-8859-1"
FAIL Name "windows-1250" has label "x-cp1250" (inputEncoding) assert_equals: expected "WINDOWS-1250" but got "ISO-8859-1"
FAIL Name "windows-1251" has label "cp1251" (characterSet) assert_equals: expected "WINDOWS-1251" but got "ISO-8859-1"
FAIL Name "windows-1251" has label "cp1251" (inputEncoding) assert_equals: expected "WINDOWS-1251" but got "ISO-8859-1"
FAIL Name "windows-1251" has label "windows-1251" (characterSet) assert_equals: expected "WINDOWS-1251" but got "ISO-8859-1"
FAIL Name "windows-1251" has label "windows-1251" (inputEncoding) assert_equals: expected "WINDOWS-1251" but got "ISO-8859-1"
FAIL Name "windows-1251" has label "x-cp1251" (characterSet) assert_equals: expected "WINDOWS-1251" but got "ISO-8859-1"
FAIL Name "windows-1251" has label "x-cp1251" (inputEncoding) assert_equals: expected "WINDOWS-1251" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ansi_x3.4-1968" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ansi_x3.4-1968" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ascii" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ascii" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "cp1252" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "cp1252" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "cp819" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "cp819" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "csisolatin1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "csisolatin1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ibm819" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "ibm819" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-8859-1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-8859-1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-ir-100" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso-ir-100" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso8859-1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso8859-1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso88591" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso88591" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1:1987" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "iso_8859-1:1987" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "l1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "l1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "latin1" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "latin1" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "us-ascii" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "us-ascii" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "windows-1252" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "windows-1252" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "x-cp1252" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "x-cp1252" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "x-user-defined" (characterSet) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1252" has label "x-user-defined" (inputEncoding) assert_equals: expected "WINDOWS-1252" but got "ISO-8859-1"
FAIL Name "windows-1253" has label "cp1253" (characterSet) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1253" has label "cp1253" (inputEncoding) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1253" has label "windows-1253" (characterSet) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1253" has label "windows-1253" (inputEncoding) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1253" has label "x-cp1253" (characterSet) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1253" has label "x-cp1253" (inputEncoding) assert_equals: expected "WINDOWS-1253" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "cp1254" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "cp1254" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "csisolatin5" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "csisolatin5" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso-8859-9" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso-8859-9" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso-ir-148" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso-ir-148" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso8859-9" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso8859-9" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso88599" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso88599" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso_8859-9" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso_8859-9" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso_8859-9:1989" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "iso_8859-9:1989" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "l5" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "l5" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "latin5" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "latin5" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "windows-1254" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "windows-1254" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "x-cp1254" (characterSet) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1254" has label "x-cp1254" (inputEncoding) assert_equals: expected "WINDOWS-1254" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "cp1255" (characterSet) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "cp1255" (inputEncoding) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "windows-1255" (characterSet) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "windows-1255" (inputEncoding) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "x-cp1255" (characterSet) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1255" has label "x-cp1255" (inputEncoding) assert_equals: expected "WINDOWS-1255" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "cp1256" (characterSet) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "cp1256" (inputEncoding) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "windows-1256" (characterSet) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "windows-1256" (inputEncoding) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "x-cp1256" (characterSet) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1256" has label "x-cp1256" (inputEncoding) assert_equals: expected "WINDOWS-1256" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "cp1257" (characterSet) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "cp1257" (inputEncoding) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "windows-1257" (characterSet) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "windows-1257" (inputEncoding) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "x-cp1257" (characterSet) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1257" has label "x-cp1257" (inputEncoding) assert_equals: expected "WINDOWS-1257" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "cp1258" (characterSet) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "cp1258" (inputEncoding) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "windows-1258" (characterSet) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "windows-1258" (inputEncoding) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "x-cp1258" (characterSet) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "windows-1258" has label "x-cp1258" (inputEncoding) assert_equals: expected "WINDOWS-1258" but got "ISO-8859-1"
FAIL Name "x-mac-cyrillic" has label "x-mac-cyrillic" (characterSet) assert_equals: expected "X-MAC-CYRILLIC" but got "ISO-8859-1"
FAIL Name "x-mac-cyrillic" has label "x-mac-cyrillic" (inputEncoding) assert_equals: expected "X-MAC-CYRILLIC" but got "ISO-8859-1"
FAIL Name "x-mac-cyrillic" has label "x-mac-ukrainian" (characterSet) assert_equals: expected "X-MAC-CYRILLIC" but got "ISO-8859-1"
FAIL Name "x-mac-cyrillic" has label "x-mac-ukrainian" (inputEncoding) assert_equals: expected "X-MAC-CYRILLIC" but got "ISO-8859-1"
FAIL Name "gbk" has label "chinese" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "chinese" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "csgb2312" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "csgb2312" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "csiso58gb231280" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "csiso58gb231280" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gb2312" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gb2312" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gb_2312" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gb_2312" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gb_2312-80" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gb_2312-80" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gbk" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "gbk" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "iso-ir-58" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "iso-ir-58" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "x-gbk" (characterSet) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gbk" has label "x-gbk" (inputEncoding) assert_equals: expected "GBK" but got "ISO-8859-1"
FAIL Name "gb18030" has label "gb18030" (characterSet) assert_equals: expected "GB18030" but got "ISO-8859-1"
FAIL Name "gb18030" has label "gb18030" (inputEncoding) assert_equals: expected "GB18030" but got "ISO-8859-1"
FAIL Name "hz-gb-2312" has label "hz-gb-2312" (characterSet) assert_equals: expected "HZ-GB-2312" but got "ISO-8859-1"
FAIL Name "hz-gb-2312" has label "hz-gb-2312" (inputEncoding) assert_equals: expected "HZ-GB-2312" but got "ISO-8859-1"
FAIL Name "big5" has label "big5" (characterSet) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "big5" (inputEncoding) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "big5-hkscs" (characterSet) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "big5-hkscs" (inputEncoding) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "cn-big5" (characterSet) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "cn-big5" (inputEncoding) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "csbig5" (characterSet) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "csbig5" (inputEncoding) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "x-x-big5" (characterSet) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "big5" has label "x-x-big5" (inputEncoding) assert_equals: expected "Big5" but got "ISO-8859-1"
FAIL Name "euc-jp" has label "cseucpkdfmtjapanese" (characterSet) assert_equals: expected "EUC-JP" but got "ISO-8859-1"
FAIL Name "euc-jp" has label "cseucpkdfmtjapanese" (inputEncoding) assert_equals: expected "EUC-JP" but got "ISO-8859-1"
FAIL Name "euc-jp" has label "euc-jp" (characterSet) assert_equals: expected "EUC-JP" but got "ISO-8859-1"
FAIL Name "euc-jp" has label "euc-jp" (inputEncoding) assert_equals: expected "EUC-JP" but got "ISO-8859-1"
FAIL Name "euc-jp" has label "x-euc-jp" (characterSet) assert_equals: expected "EUC-JP" but got "ISO-8859-1"
FAIL Name "euc-jp" has label "x-euc-jp" (inputEncoding) assert_equals: expected "EUC-JP" but got "ISO-8859-1"
FAIL Name "iso-2022-jp" has label "csiso2022jp" (characterSet) assert_equals: expected "ISO-2022-JP" but got "ISO-8859-1"
FAIL Name "iso-2022-jp" has label "csiso2022jp" (inputEncoding) assert_equals: expected "ISO-2022-JP" but got "ISO-8859-1"
FAIL Name "iso-2022-jp" has label "iso-2022-jp" (characterSet) assert_equals: expected "ISO-2022-JP" but got "ISO-8859-1"
FAIL Name "iso-2022-jp" has label "iso-2022-jp" (inputEncoding) assert_equals: expected "ISO-2022-JP" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "csshiftjis" (characterSet) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "csshiftjis" (inputEncoding) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "ms_kanji" (characterSet) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "ms_kanji" (inputEncoding) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "shift-jis" (characterSet) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "shift-jis" (inputEncoding) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "shift_jis" (characterSet) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "shift_jis" (inputEncoding) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "sjis" (characterSet) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "sjis" (inputEncoding) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "windows-31j" (characterSet) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "windows-31j" (inputEncoding) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "x-sjis" (characterSet) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "shift_jis" has label "x-sjis" (inputEncoding) assert_equals: expected "Shift_JIS" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "cseuckr" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "cseuckr" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "csksc56011987" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "csksc56011987" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "euc-kr" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "euc-kr" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "iso-ir-149" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "iso-ir-149" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "korean" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "korean" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "ks_c_5601-1987" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "ks_c_5601-1987" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "ks_c_5601-1989" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "ks_c_5601-1989" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "ksc5601" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "ksc5601" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "ksc_5601" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "ksc_5601" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "windows-949" (characterSet) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "euc-kr" has label "windows-949" (inputEncoding) assert_equals: expected "EUC-KR" but got "ISO-8859-1"
FAIL Name "replacement" has label "csiso2022kr" (characterSet) assert_equals: expected "REPLACEMENT" but got "ISO-8859-1"
FAIL Name "replacement" has label "csiso2022kr" (inputEncoding) assert_equals: expected "REPLACEMENT" but got "ISO-8859-1"
FAIL Name "replacement" has label "iso-2022-cn" (characterSet) assert_equals: expected "REPLACEMENT" but got "ISO-8859-1"
FAIL Name "replacement" has label "iso-2022-cn" (inputEncoding) assert_equals: expected "REPLACEMENT" but got "ISO-8859-1"
FAIL Name "replacement" has label "iso-2022-cn-ext" (characterSet) assert_equals: expected "REPLACEMENT" but got "ISO-8859-1"
FAIL Name "replacement" has label "iso-2022-cn-ext" (inputEncoding) assert_equals: expected "REPLACEMENT" but got "ISO-8859-1"
FAIL Name "replacement" has label "iso-2022-kr" (characterSet) assert_equals: expected "REPLACEMENT" but got "ISO-8859-1"
FAIL Name "replacement" has label "iso-2022-kr" (inputEncoding) assert_equals: expected "REPLACEMENT" but got "ISO-8859-1"
